refactor: (#66) 단축키 사용 로직 분리
